# internal/book/model.go
package book

// Book represents one row in the Book table.
type Book struct {
	ID       string  `json:"id" db:"id"`
	PubID    int     `json:"pub_id" db:"pub_id"`
	Title    string  `json:"title" db:"title"`
	Price    float64 `json:"price" db:"price"`
	Category *string `json:"category,omitempty" db:"category"`
	Quantity int     `json:"quantity" db:"quantity"`
	Format   *string `json:"b_format,omitempty" db:"b_format"`
	ProdYear int     `json:"prod_year" db:"prod_year"`
	FileSize *int    `json:"filesize,omitempty" db:"filesize"`
}

// CreateRequest contains fields accepted when creating a book.
// The ID is generated by the service.
type CreateRequest struct {
	PubID    int     `json:"pub_id" binding:"required"`
	Title    string  `json:"title" binding:"required"`
	Price    float64 `json:"price" binding:"required"`
	Category *string `json:"category,omitempty"`
	Quantity int     `json:"quantity" binding:"required"`
	Format   *string `json:"b_format,omitempty"`
	ProdYear int     `json:"prod_year" binding:"required"`
	FileSize *int    `json:"filesize,omitempty"`
}

type UpdateRequest = CreateRequest

func (r CreateRequest) Book() Book {
	return Book{
		PubID:    r.PubID,
		Title:    r.Title,
		Price:    r.Price,
		Category: r.Category,
		Quantity: r.Quantity,
		Format:   r.Format,
		ProdYear: r.ProdYear,
		FileSize: r.FileSize,
	}
}